PID controller implemented but not working

+++ Atmel Workspace/Due_robot/Due_robot/src/main.c
@@ -16,8 +16,8 @@ xSemaphoreHandle sync = (xSemaphoreHandle) 1;
 
 uint16_t cha_reading = 0;
 uint16_t chb_reading = 0;
-uint32_t cha_setpoint;
-uint32_t chb_setpoint;
+float cha_setpoint;
+float chb_setpoint;
 uint16_t new_value = 0;
